Added pre-compilation of JSX
To avoid having to transform JSX in-browser, added a build step that
will eventually pre-compile all resources (LESS, JSX).

Caveats:

* Just does JSX atm, but can be updated to compile LESS as well.
* Still need the JSXTransformer for inline JSX, would be best to move
  that out to a separate file.
* Would also be nice to perform minification as well.
