+++ codalab/apps/web/.gitignore
@@ -1,0 +1,3 @@
+node_modules
+*.swp
+static/dist
