+++ codalab/apps/web/static/js/bundle/bundle_interface.jsx
@@ -2,7 +2,6 @@
 TODO: replace this with WorksheetSidePanel.
 */
 
-/** @jsx React.DOM */
 var Bundle = React.createClass({
     getInitialState: function(){
         return {
