+++ codalab/apps/web/static/js/worksheet/bundle_uploader.jsx
@@ -1,4 +1,3 @@
-/** @jsx React.DOM */
 
 var BundleUploader = React.createClass({
   getInitialState: function() {
