+++ codalab/apps/web/static/js/worksheet/graph_item_interface.jsx
@@ -1,4 +1,3 @@
-/** @jsx React.DOM */
 
 // Display a graph over the different bundles.
 var GraphItem = React.createClass({
