+++ codalab/apps/web/static/js/worksheet/html_item_interface.jsx
@@ -1,12 +1,11 @@
-/** @jsx React.DOM */
 
-// Display a worksheet item representing the file contents of a bundle.
-var ContentsItem = React.createClass({
+// Display a worksheet item which is the HTML file in a bundle.
+var HTMLItem = React.createClass({
     mixins: [CheckboxMixin, GoToBundleMixin],
-    getInitialState: function(){
+    getInitialState: function() {
         return {};
     },
-    handleClick: function(event){
+    handleClick: function(event) {
         this.props.setFocus(this.props.focusIndex, 0);
     },
 
@@ -15,14 +14,13 @@ var ContentsItem = React.createClass({
     },
 
     render: function() {
-        var className = 'type-contents' + (this.props.focused ? ' focused' : '');
-        var contents = this.props.item.interpreted.join('');
+        var className = 'type-html' + (this.props.focused ? ' focused' : '');
+        var contents = html_sanitize(this.props.item.interpreted.join(''));
         return(
             <div className="ws-item" onClick={this.handleClick}>
                 <div className={className} ref={this.props.item.ref}>
-                    <blockquote>
-                        <p>{contents}</p>
-                    </blockquote>
+                    <div className="html-bundle" dangerouslySetInnerHTML={{__html: contents}}>
+                    </div>
                 </div>
             </div>
         );
