+++ codalab/apps/web/static/js/worksheet/image_item_interface.jsx
@@ -1,4 +1,3 @@
-/** @jsx React.DOM */
 
 // Display a worksheet item which is an image file in a bundle.
 var ImageItem = React.createClass({
