+++ codalab/apps/web/static/js/worksheet/ws_mixins.jsx
@@ -1,4 +1,3 @@
-/** @jsx React.DOM */
 
 var CheckboxMixin = {
     handleCheck: function(event){
