+++ codalab/apps/web/templates/web/bundles/detail.html
@@ -25,9 +25,8 @@
 {% block jsincludes %}
     <script src="{{ STATIC_URL }}js/vendor/react-0.12.2.js"></script>
     {# TODO: remove JSX and move to deploy processing of JSX #}
-    <script src="{{ STATIC_URL }}js/vendor/JSXTransformer-0.12.2.js"></script>
-    <script src="{{ STATIC_URL }}js/worksheet/worksheet_utils.js"></script>
-    <script type="text/jsx" src="{{ STATIC_URL }}js/bundle/bundle_interface.js"></script>
+    <script src="{{ STATIC_URL }}dist/worksheet_utils.js"></script>
+    <script src="{{ STATIC_URL }}dist/bundle_interface.js"></script>
 <script>
 //
 </script>
