+++ codalab/apps/web/templates/web/worksheets/detail.html
@@ -117,32 +117,32 @@
     <script src="{{ STATIC_URL }}js/vendor/jquery.terminal-0.8.8_custom.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-mousewheel/3.1.12/jquery.mousewheel.js"></script>
 
-    {# TODO: remove JSX and move to deploy processing of JSX #}
-    <script src="{{ STATIC_URL }}js/vendor/JSXTransformer-0.12.2.js"></script>
+    <script src="{{ STATIC_URL }}dist/worksheet_utils.js"></script>
+    <script src="{{ STATIC_URL }}dist/worksheet_content.js"></script>
+    <script src="{{ STATIC_URL }}dist/ws_mousetrap_fork.js"></script>
 
-    <script src="{{ STATIC_URL }}js/worksheet/worksheet_utils.js"></script>
-    <script src="{{ STATIC_URL }}js/worksheet/worksheet_content.js"></script>
-    <script src="{{ STATIC_URL }}js/worksheet/ws_mousetrap_fork.js"></script>
-
-    <script type="text/jsx" src="{{ STATIC_URL }}js/worksheet/ws_mixins.js"></script>
+    <script src="{{ STATIC_URL }}dist/ws_mixins.js"></script>
     {# ------- JSX FILES  ------- #}
-    <script type="text/jsx" src="{{ STATIC_URL }}js/worksheet/worksheet_action_bar.js"></script>
-    <script type="text/jsx" src="{{ STATIC_URL }}js/worksheet/worksheet_side_panel.js"></script>
+    <script src="{{ STATIC_URL }}dist/worksheet_action_bar.js"></script>
+    <script src="{{ STATIC_URL }}dist/worksheet_side_panel.js"></script>
     {# bundle type renders/actions #}
-    <script type="text/jsx" src="{{ STATIC_URL }}js/worksheet/contents_item_interface.js"></script>
-    <script type="text/jsx" src="{{ STATIC_URL }}js/worksheet/html_item_interface.js"></script>
-    <script type="text/jsx" src="{{ STATIC_URL }}js/worksheet/image_item_interface.js"></script>
-    <script type="text/jsx" src="{{ STATIC_URL }}js/worksheet/graph_item_interface.js"></script>
-    <script type="text/jsx" src="{{ STATIC_URL }}js/worksheet/markdown_item_interface.js"></script>
-    <script type="text/jsx" src="{{ STATIC_URL }}js/worksheet/record_item_interface.js"></script>
-    <script type="text/jsx" src="{{ STATIC_URL }}js/worksheet/table_item_interface.js"></script>
-    <script type="text/jsx" src="{{ STATIC_URL }}js/worksheet/worksheet_item_interface.js"></script>
-    <script type="text/jsx" src="{{ STATIC_URL }}js/worksheet/bundle_uploader.js"></script>
-    <script type="text/jsx" src="{{ STATIC_URL }}js/worksheet/editable_field.js"></script>
+    <script src="{{ STATIC_URL }}dist/contents_item_interface.js"></script>
+    <script src="{{ STATIC_URL }}dist/html_item_interface.js"></script>
+    <script src="{{ STATIC_URL }}dist/image_item_interface.js"></script>
+    <script src="{{ STATIC_URL }}dist/graph_item_interface.js"></script>
+    <script src="{{ STATIC_URL }}dist/markdown_item_interface.js"></script>
+    <script src="{{ STATIC_URL }}dist/record_item_interface.js"></script>
+    <script src="{{ STATIC_URL }}dist/table_item_interface.js"></script>
+    <script src="{{ STATIC_URL }}dist/worksheet_item_interface.js"></script>
+    <script src="{{ STATIC_URL }}dist/bundle_uploader.js"></script>
+    <script src="{{ STATIC_URL }}dist/editable_field.js"></script>
     {# main entrance point for the worksheet and it's items/bundles #}
-    <script type="text/jsx" src="{{ STATIC_URL }}js/worksheet/worksheet_items.js"></script>
-    <script type="text/jsx" src="{{ STATIC_URL }}js/worksheet/worksheet_interface.js"></script>
+    <script src="{{ STATIC_URL }}dist/worksheet_items.js"></script>
+    <script src="{{ STATIC_URL }}dist/worksheet_interface.js"></script>
     {# ------- END JSX FILES  ------- #}
+
+    {# ------- Require JSX Transformer for the inline JSX ------- #}
+    <script src="{{ STATIC_URL }}js/vendor/JSXTransformer-0.12.2.js"></script>
 
 <script type="text/jsx">
 MathJax.Hub.Config({
